+++ src/main/java/com/command/toyvillage_server/domain/app/work_log/service/WorkLogTemplateQueryListService.java
@@ -1,39 +1,17 @@
 package com.command.toyvillage_server.domain.app.work_log.service;
 
-import com.command.toyvillage_server.domain.app.work_log.domain.WorkLogTemplate;
-import com.command.toyvillage_server.domain.app.work_log.domain.WorkLogTemplateQuestion;
 import com.command.toyvillage_server.domain.app.work_log.domain.repository.WorkLogTemplateQuestionRepository;
 import com.command.toyvillage_server.domain.app.work_log.domain.repository.WorkLogTemplateRepository;
-import com.command.toyvillage_server.domain.app.work_log.exception.WorkLogTemplateNotFoundException;
-import com.command.toyvillage_server.domain.app.work_log.presentation.dto.response.WorkLogTemplateQuestionResponse;
-import com.command.toyvillage_server.domain.app.work_log.presentation.dto.response.WorkLogTemplateResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.stream.Collectors;
-
 @Service
 @RequiredArgsConstructor
 public class WorkLogTemplateQueryListService {
-    private final WorkLogTemplateQuestionRepository workLogTemplateQuestionRepository;
     private final WorkLogTemplateRepository workLogTemplateRepository;
+    private final WorkLogTemplateQuestionRepository workLogTemplateQuestionRepository;
 
     @Transactional(readOnly = true)
-    public WorkLogTemplateResponse execute(Long templateId) {
-        WorkLogTemplate workLogTemplate = workLogTemplateRepository.findById(templateId)
-            .orElseThrow(() -> WorkLogTemplateNotFoundException.EXCEPTION);
-
-        List<WorkLogTemplateQuestion> workLogTemplateQuestions = workLogTemplateQuestionRepository.findByWorkLogTemplateId(workLogTemplate.getId());
-
-        List<WorkLogTemplateQuestionResponse> workLogTemplateQuestionResponse = workLogTemplateQuestions.stream()
-            .map(WorkLogTemplateQuestionResponse::from)
-            .toList();
-
-        return WorkLogTemplateResponse.of(
-            workLogTemplate.getTemplateTitle(),
-            workLogTemplateQuestionResponse
-        );
-    }
+    public
 }
